+++ rMall/Domain/src/main/java/ecommerce/rmall/domain/Credential.java
@@ -1,0 +1,48 @@
+package ecommerce.rmall.domain;
+
+import java.util.Date;
+
+import javax.xml.bind.annotation.XmlRootElement;
+
+import org.codehaus.jackson.annotate.JsonIgnore;
+
+/***
+ * 认证信息
+ * @author martin
+ *
+ */
+@XmlRootElement (name = "Credential")
+public class Credential{
+	
+	@JsonIgnore
+	private String username;
+	@JsonIgnore
+	private String password;
+	private String sessionKey;
+	private Date expireTime;
+	
+	public String getUsername() {
+		return username;
+	}
+	public void setUsername(String username) {
+		this.username = username;
+	}
+	public String getPassword() {
+		return password;
+	}
+	public void setPassword(String password) {
+		this.password = password;
+	}
+	public String getSessionKey() {
+		return sessionKey;
+	}
+	public void setSessionKey(String sessionKey) {
+		this.sessionKey = sessionKey;
+	}
+	public Date getExpireTime() {
+		return expireTime;
+	}
+	public void setExpireTime(Date expireTime) {
+		this.expireTime = expireTime;
+	}
+}
