+++ rMall/Domain/src/main/java/ecommerce/rmall/domain/Station.java
@@ -7,44 +7,18 @@ import javax.xml.bind.annotation.XmlRootElement;
 @XmlRootElement (name = "Station")
 public class Station {
 	
-	static public class Credential{
-		
-		public String username;
-		public String password;
-		public String sessionKey;
-		public Date expireTime;
-		
-		public String getUsername() {
-			return username;
-		}
-		public void setUsername(String username) {
-			this.username = username;
-		}
-		public String getPassword() {
-			return password;
-		}
-		public void setPassword(String password) {
-			this.password = password;
-		}
-		public String getSessionKey() {
-			return sessionKey;
-		}
-		public void setSessionKey(String sessionKey) {
-			this.sessionKey = sessionKey;
-		}
-		public Date getExpireTime() {
-			return expireTime;
-		}
-		public void setExpireTime(Date expireTime) {
-			this.expireTime = expireTime;
-		}
-		
-	}
-	
 	private int id;
 	private Date createDate;
 	private String name;
+	private String city;
 	private Credential credential;
+	
+	public String getCity() {
+		return city;
+	}
+	public void setCity(String city) {
+		this.city = city;
+	}
 	
 	public int getId() {
 		return id;
