+++ rMall/mallService/pom.xml
@@ -16,6 +16,18 @@
 	<build>
 		<finalName>mallService</finalName>
 		<plugins>
+<plugin>
+    <groupId>org.codehaus.mojo</groupId> 
+    <artifactId>tomcat-maven-plugin</artifactId> 
+    <version>1.1</version> 
+    <configuration>
+        <url>http://localhost:8080/manager/text</url>
+        <server>test</server>
+        <username>tomcat</username>
+        <password>tomcat</password>  
+    </configuration>
+</plugin>
+			  
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-resources-plugin</artifactId>
